+++ app/src/main/java/com/bluecat94/taskalert/ui/TaskDetailActivity.java
@@ -26,14 +26,14 @@ public class TaskDetailActivity extends AppCompatActivity {
         Intent intent = getIntent();
         String title = intent.getStringExtra(TasksContract.TaskEntry.COLUMN_TITLE);
         String description = intent.getStringExtra(TasksContract.TaskEntry.COLUMN_DESCRIPTION);
-        long longitude = intent.getLongExtra(TasksContract.TaskEntry.COLUMN_LONGITUDE, 0);
-        long lat = intent.getLongExtra(TasksContract.TaskEntry.COLUMN_LATITTUDE, 0);
+        double longitude = intent.getDoubleExtra(TasksContract.TaskEntry.COLUMN_LONGITUDE, 0);
+        double lat = intent.getDoubleExtra(TasksContract.TaskEntry.COLUMN_LATITTUDE, 0);
 
         Bundle bundle = new Bundle();
         bundle.putString(TasksContract.TaskEntry.COLUMN_TITLE, title);
         bundle.putString(TasksContract.TaskEntry.COLUMN_DESCRIPTION, description);
-        bundle.putLong(TasksContract.TaskEntry.COLUMN_LATITTUDE, lat);
-        bundle.putLong(TasksContract.TaskEntry.COLUMN_LONGITUDE, longitude);
+        bundle.putDouble(TasksContract.TaskEntry.COLUMN_LATITTUDE, lat);
+        bundle.putDouble(TasksContract.TaskEntry.COLUMN_LONGITUDE, longitude);
 
         FragmentManager fragmentManager = getSupportFragmentManager();
         FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
